requirements.txt
# Core ML
faster-whisper>=1.0.0
torch>=2.1.0
torchaudio>=2.1.0

# Speaker Diarization
# 本地方案（推薦，無需 token）
simple-diarizer>=0.0.6
resemblyzer>=0.1.1
spectralcluster>=0.1.0
# 高精度方案（可選，需一次性 HF token 下載）
pyannote.audio>=3.1.0

# Audio Recording
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0

# UI
PyQt6>=6.6.0

# Storage & Export
sqlalchemy>=2.0.0
python-docx>=1.1.0
reportlab>=4.0.0
fpdf2>=2.7.0

# Local LLM
ollama>=0.2.0

# Utils
requests>=2.31.0
python-dotenv>=1.0.0
